requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.0
bootstrap4==0.1.0
certifi==2022.9.14
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
comtypes==1.1.14
constantly==15.1.0
crayons==0.4.0
crispy-bootstrap5==0.7
cryptography==41.0.2
daphne==4.0.0
dj-database-url==1.3.0
Django==4.1.7
django-bootstrap4==22.3
django-crispy-forms==2.0
django-debug-toolbar==4.1.0
Django-Gtts==0.4
django-heroku==0.3.1
django-redis==5.3.0
django-redis-sessions==0.6.2
django-redis-views==0.2.2
djangorestframework==3.14.0
googletrans==4.0.0rc1
gTTS==2.2.4
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.9.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==22.10.0
msgpack==1.0.5
mysql-connector-python==8.0.30
mysqlclient==2.1.1
numpy==1.24.3
opencv-python==4.7.0.72
pickle-mixin==1.0.2
piexif==1.1.3
Pillow==9.5.0
protobuf==3.20.1
psycopg2==2.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
pypiwin32==223
pyttsx3==2.90
pytz==2023.3
pywin32==304
redis==4.6.0
requests==2.28.1
rfc3986==1.5.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
sqlparse==0.4.2
stegano==0.11.2
stripe==5.5.0
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2022.1
urllib3==1.26.12
whitenoise==6.4.0
zope.interface==6.0
